more partial results
Inner Join seems to be working, aside from some column projection issues that need cleaning up.

Performance is good compared to the worst case performance of the old version.
