+++ core/src/test/java/tech/tablesaw/joining/DataFrameJoinerPerformanceTest.java
@@ -71,8 +71,7 @@ public class DataFrameJoinerPerformanceTest {
     Table orders = createOrdersTable(numberOrders, 2_000);
     addFillerColumn(orders, 5, "order");
 
-    assertTimeout(
-        Duration.ofSeconds(1), () -> orders.joinOn("customerId").leftOuterJoin(customers));
+    assertTimeout(Duration.ofSeconds(1), () -> orders.joinOn("customerId").leftOuter(customers));
   }
 
   @Test
